requirements.txt
request
json
dotenv
os
pandas
line-bot-sdk
matplotlib
flask
re
io
openai
random
sqlite3
datetime
qrcode